move all scroll logic into context
