+++ packages/web/app/contexts/LocationBookings.jsx
@@ -1,8 +1,15 @@
 import React, { createContext, useContext, useEffect, useState } from 'react';
+import { isThisMonth, startOfToday } from 'date-fns';
 
 import { scrollToCell } from '../views/scheduling/locationBookings/utils';
+import { firstDisplayedDayId, thisWeekId } from '../views/scheduling/locationBookings/LocationBookingsCalendarHeader';
 
 const LocationBookingsContext = createContext(null);
+
+const scrollToThisWeek = () =>
+  document.getElementById(thisWeekId)?.scrollIntoView({ inline: 'start' });
+const scrollToBeginning = () =>
+  document.getElementById(firstDisplayedDayId)?.scrollIntoView({ inline: 'start' });
 
 export const LocationBookingsContextProvider = ({ children }) => {
   const [filters, setFilters] = useState({
@@ -15,17 +22,23 @@ export const LocationBookingsContextProvider = ({ children }) => {
   const [selectedCell, setSelectedCell] = useState({
     locationId: null,
     date: null,
-  }); 
+  });
 
+  const [monthOf, setMonthOf] = useState(startOfToday());
+
+  // Calendar Scroll logic
   useEffect(() => {
     if (selectedCell.locationId && selectedCell.date) {
       scrollToCell(selectedCell);
+      return;
     }
-  }, [selectedCell]);
+    (isThisMonth(monthOf) ? scrollToThisWeek : scrollToBeginning)();
+  }, [selectedCell, monthOf]);
+
 
   return (
     <LocationBookingsContext.Provider
-      value={{ filters, setFilters, selectedCell, setSelectedCell }}
+      value={{ filters, setFilters, selectedCell, setSelectedCell, monthOf, setMonthOf }}
     >
       {children}
     </LocationBookingsContext.Provider>
